Recactoring, one service to handle all logc
